+++ src/device/eltako_fsr61/fsr61_actor.py
@@ -126,4 +126,4 @@ class Fsr61Actor(SceneActor, CheckCyclicTask):
 
     @property
     def _randomized_refresh_rate(self) -> int:
-        return self.DEFAULT_REFRESH_RATE + random.randint(0, self.DEFAULT_REFRESH_RATE * 0.1)
+        return self.DEFAULT_REFRESH_RATE + random.randint(0, int(self.DEFAULT_REFRESH_RATE * 0.1))
